fix(W0-2): Gemini round 1+2 PR #4 全採納（LongPressData / UInt / init require / Fakes 唯讀）
【Gemini round 1+2 PR #4 9 條 inline 全採納】
- KeyData.kt: 新增 LongPressData 封裝 label+action；KeyData 用單一可空欄位
  longPress: LongPressData? 取代 longPressLabel/longPressAction 兩個鬆散欄位。
  型別系統一次解掉「同生同滅」不變量，不再需要 runtime require
- KeyboardTheme.kt: KeyboardColors 6 個顏色欄位 Long → UInt。避免 sign
  extension（0xFFFFFFFFL = -1，toULong() 高位填 1 會讓 Compose Color 誤判
  color space）。UInt.toULong() 高位保證 0。完整 KDoc 解釋
- InputState.kt: 加 init { require ... } 強制驗證 cursorIndex：candidates
  空時必須 0，非空時必須在 indices 範圍內。違反就 IllegalArgumentException
- FakeImeContextProvider.kt: deleteCalls / sentKeys 從 public MutableList
  改為 private _deleteCalls / _sentKeys + 公開 List 唯讀 getter
- FakeZhuyinDecoder.kt: committedHistory 同上改唯讀封裝
- FakeKeyboardTheme.kt: DEFAULT_COLORS 改用 UInt 字面常數
- SerializationRoundTripTest.kt: 更新 KeyData 測試用 LongPressData、
  KeyboardColors 測試改 UInt
- FakeKeyboardThemeTest.kt: ushr → shr（UInt 對應 op）、0xFFL → 0xFFu

驗收：:common:build :common:test :common:ktfmtCheck 全綠（11s, cc reused）
所有 test pass。Wire format：KeyData 多一層 longPress 物件；KeyboardColors
JSON 是 unsigned number。

+++ common/src/main/kotlin/com/bopomofobruce/common/KeyboardTheme.kt
@@ -3,10 +3,16 @@ package com.bopomofobruce.common
 import kotlinx.serialization.Serializable
 
 /**
- * 鍵盤主題的顏色集合。所有顏色用 [Long] 存 ARGB（不是 Compose Color，避免 :common 依賴 Compose）。
+ * 鍵盤主題的顏色集合。所有顏色用 [UInt] 存 ARGB（不是 Compose Color，避免 :common 依賴 Compose）。
  *
- * 慣例：高 8 bit alpha、再 RGB，例如 `0xFFFF0000L` 是不透明紅。 :keyboards / :theme 端會把 [Long] 轉成
- * `androidx.compose.ui.graphics.Color(value.toULong())`。
+ * 為什麼用 `UInt` 不用 `Long`：
+ * - 32 位元無符號精確對應 ARGB；不會出現負值或超出 32-bit 的範圍
+ * - 避免 sign extension：例如 `0xFFFFFFFF` 在 Int/Long 是 -1，後續 `toULong()` 高位會被填 1， Compose `Color(value:
+ *   ULong)` 會誤判 color space。`UInt.toULong()` 高位保證 0
+ *
+ * 慣例：高 8 bit alpha、再 RGB，例如 `0xFFFF0000u` 是不透明紅。 :keyboards / :theme 端會把 [UInt] 轉成
+ * `androidx.compose.ui.graphics.Color(value.toULong() shl 32)` 或 `Color(value.toInt())`（看 Compose
+ * 版本）。
  * - [background]：整個 IME panel 背景。
  * - [keyFill]：按鍵填色（normal state）。
  * - [keyText]：按鍵文字色。
@@ -16,12 +22,12 @@ import kotlinx.serialization.Serializable
  */
 @Serializable
 data class KeyboardColors(
-    val background: Long,
-    val keyFill: Long,
-    val keyText: Long,
-    val keyAccent: Long,
-    val candidateText: Long,
-    val candidateHighlight: Long,
+    val background: UInt,
+    val keyFill: UInt,
+    val keyText: UInt,
+    val keyAccent: UInt,
+    val candidateText: UInt,
+    val candidateHighlight: UInt,
 )
 
 /**
